SERVER/requirements.txt
Flask
Flask-RESTful
Flask-CORS
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
SQLAlchemy-Utils
passlib
python-jose
PyJWT